Split design module into design + devices modules.
