+++ beamz/devices/signals.py
